Crate zaino_serve

Holds a gRPC server capable of servicing clients over TCP.

server::ingestor has been built so that other ingestors may be added that use different transport protocols (Nym, TOR).

Also holds rust implementations of the LightWallet gRPC Service (CompactTxStreamerServer).

Modules

rpc

Lightwallet service RPC implementations.

server

Zaino’s gRPC server implementation.